requirements.txt
Pillow
apify-client
insightface
opencv-python
ruff
python-dotenv
requests
unidecode
numpy
onnxruntime-silicon
